Multiple options for Swipe-to-close on Android Notification
Now the app developer can specify whether the notification will always
close on swipe, only close if the audio is paused, or will never close
when swiped.  In addition, a new control was added to provide a
callback for when it is closed via swipe, so that the app developer can
choose to terminate the audio or do nothing.